+++ .firebaserc
@@ -1,0 +1,5 @@
+{
+  "projects": {
+    "default": "ecommercecust-12344"
+  }
+}
