+++ src/config/axios.js
@@ -1,7 +1,7 @@
 import axios from 'axios'
 
 const instance = axios.create({
-  baseURL: 'http://localhost:3000'
+  baseURL: 'https://fast-reaches-72869.herokuapp.com/'
 })
 
 export default instance
